# api/openapi.yaml
openapi: 3.1.0
info:
  title: Cairn API
  version: 0.1.0
  description: Versioned API shared by Cairn desktop, PWA, and future mobile clients.
  license:
    name: GPL-3.0-only
servers:
  - url: http://127.0.0.1:7381
paths:
...
  /api/v1/feeds:
    get:
      operationId: listFeeds
      tags: [feeds]
      responses:
        "200":
          description: Feeds with subscription metadata.
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/Feed"
    post:
      operationId: createFeed
      tags: [feeds]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateFeedRequest"
      responses:
        "201":
          description: Feed and subscription created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Feed"
        "400":
          $ref: "#/components/responses/Problem"
  /api/v1/feeds/discover:
    post:
      operationId: discoverFeeds
      tags: [feeds]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [url]
              properties:
                url:
                  type: string
                  format: uri
      responses:
        "200":
          description: Feed candidates discovered from the URL.
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/FeedCandidate"
        "400":
          $ref: "#/components/responses/Problem"
  /api/v1/feeds/{feedId}:
    parameters:
      - $ref: "#/components/parameters/FeedId"
    get:
      operationId: getFeed
      tags: [feeds]
      responses:
        "200":
          description: Feed detail.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Feed"
        "404":
          $ref: "#/components/responses/Problem"
    patch:
      operationId: updateFeed
      tags: [feeds]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateFeedRequest"
      responses:
        "200":
          description: Updated subscription settings.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subscription"
        "404":
          $ref: "#/components/responses/Problem"
    delete:
      operationId: deleteFeed
      tags: [feeds]
      responses:
        "204":
          description: Subscription removed. Orphan cleanup is asynchronous.
        "404":
          $ref: "#/components/responses/Problem"
  /api/v1/feeds/{feedId}/refresh:
    parameters:
      - $ref: "#/components/parameters/FeedId"
    post:
      operationId: refreshFeed
      tags: [feeds, jobs]
      responses:
        "202":
          description: Refresh job accepted.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
...
  /api/v1/folders:
    get:
      operationId: listFolders
      tags: [subscriptions]
      responses:
        "200":
          description: Folder hierarchy source data.
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/Folder"
    post:
      operationId: createFolder
      tags: [subscriptions]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateFolderRequest"
      responses:
        "201":
          description: Folder created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Folder"
  /api/v1/folders/{folderId}:
    parameters:
      - $ref: "#/components/parameters/FolderId"
    patch:
      operationId: updateFolder
      tags: [subscriptions]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateFolderRequest"
      responses:
        "200":
          description: Folder updated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Folder"
        "409":
          $ref: "#/components/responses/Problem"
    delete:
      operationId: deleteFolder
      tags: [subscriptions]
      responses:
        "204":
          description: Folder deleted.
        "404":
          $ref: "#/components/responses/Problem"
  /api/v1/entries:
    get:
      operationId: listEntries
      tags: [entries]
      parameters:
        - name: cursor
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 30
        - name: feed_id
          in: query
          schema:
            type: string
            format: uuid
        - name: folder_id
          in: query
          schema:
            type: string
            format: uuid
        - name: tag_id
          in: query
          schema:
            type: string
            format: uuid
        - name: state
          in: query
          schema:
            type: string
            enum: [all, unread, starred, read_later]
        - name: query
          in: query
          schema:
            type: string
            maxLength: 500
        - name: since
          in: query
          schema:
            type: string
            format: date-time
      responses:
        "200":
          description: Cursor-paginated timeline.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EntryPage"
  /api/v1/entries/mark-read:
    post:
      operationId: markEntriesRead
      tags: [entries]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                feed_id:
                  type: string
                  format: uuid
                folder_id:
                  type: string
                  format: uuid
                tag_id:
                  type: string
                  format: uuid
                state:
                  type: string
                  enum: [all, unread, starred, read_later]
                query:
                  type: string
                  maxLength: 500
                since:
                  type: string
                  format: date-time
      responses:
        "200":
          description: Matching unread entries were marked read.
          content:
            application/json:
              schema:
                type: object
                required: [updated]
                properties:
                  updated:
                    type: integer
                    minimum: 0
  /api/v1/entries/{entryId}:
    parameters:
      - $ref: "#/components/parameters/EntryId"
    get:
      operationId: getEntry
      tags: [entries]
      responses:
        "200":
          description: Entry detail and sanitized content.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EntryDetail"
        "404":
          $ref: "#/components/responses/Problem"
  /api/v1/entries/{entryId}/state:
    parameters:
      - $ref: "#/components/parameters/EntryId"
    patch:
      operationId: updateEntryState
      tags: [entries]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateEntryStateRequest"
      responses:
        "200":
          description: Authoritative entry state.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EntryState"
  /api/v1/entries/{entryId}/readability:
    parameters:
      - $ref: "#/components/parameters/EntryId"
    post:
      operationId: fetchReadability
      tags: [entries, jobs]
      responses:
        "202":
          description: Full-text extraction job accepted.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
  /api/v1/entries/{entryId}/tags:
    parameters:
      - $ref: "#/components/parameters/EntryId"
    put:
      operationId: setEntryTags
      tags: [entries, tags]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [tag_ids]
              properties:
                tag_ids:
                  type: array
                  items:
                    type: string
                    format: uuid
      responses:
        "200":
          description: Entry tags replaced.
          content:
            application/json:
              schema:
                type: object
                required: [tag_ids]
                properties:
                  tag_ids:
                    type: array
                    items:
                      type: string
                      format: uuid
  /api/v1/tags:
    get:
      operationId: listTags
      tags: [tags]
      responses:
        "200":
          description: Profile tags.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TagList"
    post:
      operationId: createTag
      tags: [tags]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTagRequest"
      responses:
        "201":
          description: Tag created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tag"
  /api/v1/tags/{tagId}:
    delete:
      operationId: deleteTag
      tags: [tags]
      parameters:
        - $ref: "#/components/parameters/TagId"
      responses:
        "204":
          description: Tag deleted.
        "404":
          $ref: "#/components/responses/Problem"
  /api/v1/rules:
    get:
      operationId: listRules
      tags: [rules]
      responses:
        "200":
          description: Automation rules.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RuleList"
    post:
      operationId: createRule
      tags: [rules]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRuleRequest"
      responses:
        "201":
          description: Rule created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Rule"
  /api/v1/rules/{ruleId}:
    delete:
      operationId: deleteRule
      tags: [rules]
      parameters:
        - $ref: "#/components/parameters/RuleId"
      responses:
        "204":
          description: Rule deleted.
        "404":
          $ref: "#/components/responses/Problem"
  /api/v1/saved-filters:
    get:
      operationId: listSavedFilters
      tags: [filters]
      responses:
        "200":
          description: Saved timeline filters.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SavedFilterList"
    post:
      operationId: createSavedFilter
      tags: [filters]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSavedFilterRequest"
      responses:
        "201":
          description: Saved filter created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SavedFilter"
  /api/v1/saved-filters/{filterId}:
    delete:
      operationId: deleteSavedFilter
      tags: [filters]
      parameters:
        - $ref: "#/components/parameters/FilterId"
      responses:
        "204":
          description: Saved filter deleted.
        "404":
          $ref: "#/components/responses/Problem"
...
components:
  securitySchemes:
    deviceToken:
      type: http
      scheme: bearer
      bearerFormat: CairnDeviceToken
  parameters:
    FeedId:
      name: feedId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  schemas:
...
    Feed:
      type: object
      required: [id, url, title, failure_count, created_at, updated_at]
      properties:
        id:
          type: string
          format: uuid
        url:
          type: string
          format: uri
        canonical_url:
          type: [string, "null"]
          format: uri
        site_url:
          type: [string, "null"]
          format: uri
        title:
          type: string
        description:
          type: [string, "null"]
        icon_url:
          type: [string, "null"]
          format: uri
        format:
          type: [string, "null"]
          enum: [rss, atom, json, null]
        last_checked_at:
          type: [string, "null"]
          format: date-time
        last_success_at:
          type: [string, "null"]
          format: date-time
        failure_count:
          type: integer
          minimum: 0
        last_error_code:
          type: [string, "null"]
        last_error_message:
          type: [string, "null"]
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
security:
  - deviceToken: []